requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-model-utils==4.2.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.44
phonenumberslite==8.12.44
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
